[Refactor] 추천 시간 컴포넌트를 폴더 구조로 분리하고 절대 경로로 정리
RecommendedTimes 한 파일의 헤딩·리스트를 컴포넌트별 폴더로 분리하고, 미사용
컨테이너를 제거. RecommendedTime을 소비처 RecommendedTimesList 하위로 옮기고,
shared로 향하던 긴 상대 경로를 @/ alias 절대 경로로 변경
