--- src/features/event/components/detail/MainContent/DesktopSideContent/RecommendedTimes/index.tsx
@@ -1,5 +1,0 @@
-export {
-  default,
-  RecommendedTimesHeading,
-  RecommendedTimesList,
-} from './RecommendedTimes';
