+++ src/features/event/components/detail/MainContent/DesktopSideContent/RecommendedTimesHeading/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './RecommendedTimesHeading';
